Added meta selection

+++ shared/src/LgsUtils.cpp
@@ -149,7 +149,7 @@ std::string scanEscapeStr(const std::string& value) {
     return out;
 }
 
-int32_t hashString(const std::string& str) {
+uint64_t hashString(const std::string& str) {
     return std::hash<std::string_view>{}(str);
 }
